+++ .eslintrc.json
@@ -12,5 +12,5 @@
     "eslint:recommended", // use recommended configs
     "plugin:react/recommended"
   ],
-  "rules": { "react/prop-types": 0 }
+  "rules": { "react/prop-types": 0, "no-console": 0 }
 }
