+++ src/components/_index.scss
@@ -1,4 +1,3 @@
 @import './navbar';
 @import './footer';
-@import './card';
 @import './pillArticle';
