+++ src/components/navbar/_index.scss
@@ -17,10 +17,4 @@
   .logo {
     width: 100%;
   }
-
-  .page-container {
-    background-color: var(--bg);
-    transition: all 0.3s ease;
-    width: 100vw;
-  }
 }
