+++ src/layouts/_index.scss
@@ -1,0 +1,2 @@
+@import './mainLayout';
+@import './cardLayout';
