+++ src/layouts/mainLayout/_index.scss
@@ -1,0 +1,5 @@
+.page-container {
+  background-color: var(--bg);
+  transition: all 0.3s ease;
+  width: 100vw;
+}
