+++ src/layouts/mainLayout/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 
-import Navbar from "../components/navbar";
-import Footer from "../components/footer";
+import Navbar from "../../components/navbar";
+import Footer from "../../components/footer";
 
 const layout = ({ children }) => {
   return (
