+++ src/styles/index.scss
@@ -2,3 +2,4 @@
 @import '../components';
 @import '../templates';
 @import '../pages';
+@import '../layouts';
